Add Apple Silicon (MPS) inference support
Enables running the RealRestorer pipeline on macOS with Metal Performance
Shaders (MPS) without affecting the existing CUDA path. All changes are
guarded so CUDA users see no behavioral difference.

Three small fixes were required:

1. inference.py: auto-detect device (cuda > mps > cpu) and attention mode.
   The model code already auto-falls back to "torch" attention when
   flash_attn is unavailable, but surfacing the right default here keeps
   --mode honest.

2. layers.py:
   - to_cuda(): use torch.cuda > MPS > CPU instead of a hard .cuda()
     call. This helper is only invoked under cpu_offload_checkpointing
     during training, but making it portable removes an Apple Silicon
     footgun for anyone enabling that path.
   - rope(): cast intermediate scale tensor to float32 on MPS. MPS does
     not support float64 (PyTorch known issue, see e.g. diffusers
     issue #12653). The function returns .float() at the end, so the
     precision change is bounded to the omega/cos/sin intermediates and
     is numerically negligible for theta=10000, head_dim<=128.

3. pipeline_realrestorer.py: replace two hardcoded torch.bfloat16 dtypes
   in __call__ with self.transformer.dtype. The hardcoding made the
   pipeline crash whenever the model was loaded in any dtype other than
   bfloat16; using the runtime transformer dtype is correct on every
   backend (CUDA bf16/fp16 and MPS bf16).

Tested on:
- Apple M5 Max, macOS 26.4.1, PyTorch 2.8.0, transformers 4.57.3
- bf16 / 1024px / 28 steps end-to-end inference produces a valid restored
  image (~7 min on M5 Max).
- Not retested on CUDA in this environment; changes are additive and
  shouldn't regress that path, but a maintainer pass on a CUDA box would
  be appreciated.

+++ diffusers/src/diffusers/pipelines/realrestorer/pipeline_realrestorer.py
@@ -583,7 +583,7 @@ class RealRestorerPipeline(DiffusionPipeline):
             ref_images_raw = self.load_image(resized_image).to(device=device, dtype=torch.float32)
             height, width = ref_images_raw.shape[-2:]
             ref_latents_tensor = self._encode_vae_image(ref_images_raw)
-            ref_latents = self._pack_latents(ref_latents_tensor.to(device=device, dtype=torch.bfloat16))
+            ref_latents = self._pack_latents(ref_latents_tensor.to(device=device, dtype=self.transformer.dtype))
 
         if height is None or width is None:
             raise ValueError("Both height and width must be resolved before sampling.")
@@ -597,7 +597,7 @@ class RealRestorerPipeline(DiffusionPipeline):
             ),
             generator=generator,
             device=device,
-            dtype=torch.bfloat16,
+            dtype=self.transformer.dtype,
         )
         latents = self._pack_latents(noise)
